backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.44
pydantic==2.12.5
pytest==9.0.1
httpx==0.27.2
python-multipart==0.0.12
ipfshttpclient==0.7.0
web3==7.14.0
python-dotenv==1.0.0
